(bug): Fix multi-subscript skipped over closing brace
Before if you tried to do `x5` -> `x_5` and then `x_51` -> `x_{51}`
your point would be moved outside of the closing brace so if you wanted
to add more to the body of the subscript you couldn't. Now you're kept
in there and have to exit the closing brace manually.